Added full support of refresh token. If a user sends client_id and client_secret, then either refresh token is returned or an error. If there is no client_id, only access token will be returned without a refresh token.
Added Client model to represent front-end applications which have access to API. Now client check is disable in code. But if client_id is passed, it must go with client_secret and they must match the DB values (in Seed method of configuration I added react and flutter clients). It also contains how much the refresh token for this client lives and Access-control-allow-origin for each client. In addition there is field whether the client is active - if not, requests are declined from it.

Added RefreshToken model to store refresh token in DB along with Ticket. There is only 1 refresh token for a pair of a user and a client. Id is the refresh token.

Client secret, refresh token are stored in hash. Hashing algorithm is contained in Helper.cs.